+++ app/app.js
@@ -8,12 +8,10 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
 const auth = require('./src/apis/root');
-const signUp = require('./src/apis/sign-up');
 const review = require('./src/apis/review');
 const home = require('./src/apis/CircleHome');
 
 app.use('/api', auth);
-app.use('/api', signUp);
 app.use('/api/club/review', review);
 app.use('/api/club', home);
