--- app/src/apis/sign-up/index.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const express = require('express');
-
-const router = express.Router();
-const ctrl = require('./sign-up.ctrl');
-
-// 회원가입 API
-router.post('/sign-up', ctrl.process.signUp);
-
-module.exports = router;
